requirements.txt
# Steps
# 1. Create and activate venv
# 2. Install packages compiled for windows
#     - pip install packages\GDAL-3.4.2-cp39-cp39-win_amd64.whl
#     - pip install packages\Fiona-1.8.21-cp39-cp39-win_amd64.whl
#     - pip install packages\rasterio-1.2.10-cp39-cp39-win_amd64.whl
# 3. Install requirements, including pytorch for CUDA 11
#     - pip install -r requirements.txt --extra-index-url https://download.pytorch.org/whl/cu116


# Base ----------------------------------------
matplotlib>=3.2.2
numpy>=1.18.5, <1.24
opencv-python>=4.1.1
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
torch>=1.7.0,!=1.12.0
torchvision>=0.8.1,!=0.13.0
tqdm>=4.41.0
protobuf<4.21.3

# Logging -------------------------------------
tensorboard>=2.4.1
wandb

# Plotting ------------------------------------
pandas>=1.1.4
seaborn>=0.11.0

# Export --------------------------------------
# coremltools>=4.1  # CoreML export
# onnx>=1.9.0  # ONNX export
# onnx-simplifier>=0.3.6  # ONNX simplifier
# scikit-learn==0.19.2  # CoreML quantization
# tensorflow>=2.4.1  # TFLite export
# tensorflowjs>=3.9.0  # TF.js export
# openvino-dev  # OpenVINO export

# Extras --------------------------------------
ipython  # interactive notebook
psutil  # system utilization
thop  # FLOPs computation
# albumentations>=1.0.3
# pycocotools>=2.0  # COCO mAP
roboflow

# Project -------------------------------------
black[d]
ephem
gdal-ecw
geopandas
jupyter
notebook-as-pdf
pandas
pyodbc
python-dotenv
scikit-learn
seaborn
shapely
skyfield
